Authentication requests, admin functionality

+++ routers/todos.py
@@ -6,6 +6,7 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from starlette import status
 from database.db_manager import SessionLocal
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -17,6 +18,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -24,31 +26,41 @@ class TodoRequest(BaseModel):
     priority: int = Field(ge=1, le=5)
     complete: bool
 
-@router.get("/")
-async def read_all(db: db_dependency):
+#get all todos using the user id
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(user: user_dependency, db: db_dependency):
     # Add parentheses to all() to execute query
-    return db.query(Todos).all()
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo( db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).first()
     if todo_model is not None:  #if there is data in the model
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump()) #unpack the data
+async def create_todo(user: user_dependency, db: db_dependency, 
+                      todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate credentials')  
+    
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id')) #unpack the data
 
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
-                      todo_request: TodoRequest,
-                       todo_id: int = Path(gt=0)
-                       ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(
+    user: user_dependency,
+    db: db_dependency,
+    todo_request: TodoRequest,
+    todo_id: int = Path(gt=0)
+):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate credentials')  
+     
+    todo_model = db.query(Todos).filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     
@@ -62,13 +74,16 @@ async def update_todo(db: db_dependency,
     db.commit()
 
 
-    @app.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-    async def delete_todo(db: db_dependency,
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user: user_dependency,
+                      db: db_dependency,
                           todo_id: int = Path(gt=0)
                           ):
-        todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+        if user is None:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate credentials')  
+        todo_model = db.query(Todos).filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).first()
         if todo_model is None:
             raise HTTPException(status_code=404, detail="Todo not found")
-        
-        db.query(Todos).filter(Todos.id == todo_id).delete()
+
+        db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
         db.commit()
